# Conflicts:
#	libraries/tokopatch/src/main/java/com/tokopedia/tokopatch/domain/repository/PatchRepository.kt
#	libraries/tokopatch/src/main/java/com/tokopedia/tokopatch/patch/PatchExecutors.kt
#	libraries/tokopatch/src/main/java/com/tokopedia/tokopatch/patch/PatchService.kt
#	libraries/tokopatch/src/main/java/com/tokopedia/tokopatch/utils/PatchLogger.kt

+++ features/discovery/home_wishlist/src/main/java/com/tokopedia/home_wishlist/model/action/AddToCartActionData.kt
@@ -4,7 +4,7 @@ import com.tokopedia.home_wishlist.model.entity.WishlistItem
 
 data class AddToCartActionData(
         val position: Int = -1,
-        val productId: Int = -1,
+        val productId: Long = -1L,
         val cartId: String = "",
         val item: WishlistItem = WishlistItem(),
         override val message: String = "",
