+++ features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/quickedit/price/data/model/ProductEditPriceParam.kt
@@ -2,7 +2,7 @@ package com.tokopedia.product.manage.feature.quickedit.price.data.model
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
-import com.tokopedia.product.manage.feature.quickedit.common.data.model.ShopParam
+import com.tokopedia.product.manage.common.feature.quickedit.common.data.model.ShopParam
 
 data class ProductEditPriceParam(
         @SerializedName("productID")
@@ -10,7 +10,7 @@ data class ProductEditPriceParam(
         var productId: String = "",
         @SerializedName("price")
         @Expose
-        var price: Float = 0f,
+        var price: Double = 0.0,
         @SerializedName("shop")
         @Expose
         var shop: ShopParam = ShopParam()
