+++ features/user/topchat/src/main/java/com/tokopedia/topchat/chatroom/view/listener/DualAnnouncementListener.kt
@@ -4,6 +4,6 @@ package com.tokopedia.topchat.chatroom.view.listener
  * @author by nisie on 11/01/19.
  */
 interface DualAnnouncementListener{
-    fun onDualAnnouncementClicked(redirectUrl: String, attachmentId: String, blastId: Int)
+    fun onDualAnnouncementClicked(redirectUrl: String, attachmentId: String, blastId: Long)
 
 }
